# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Autenticación",
        "description": "API para la autenticación de usuarios",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/sensors": {
            "get": {
                "tags": [
                    "Sensors"
                ],
                "summary": "Obtener todos los sensores",
                "operationId": "b4af7f406a8dfe72b2804568c6eb1fd3",
                "responses": {
                    "200": {
                        "description": "Lista de sensores"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Sensors"
                ],
                "summary": "Crear un nuevo sensor",
                "operationId": "8cd7026984a4b38e0e82a3e0dfd74314",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "device_id",
                                    "type",
                                    "name"
                                ],
                                "properties": {
                                    "device_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "temperatura"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Sensor Temp 1"
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "device_id": 1,
                                "type": "temperatura",
                                "name": "Sensor Temp 1"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sensor creado exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/sensors/{id}": {
            "get": {
                "tags": [
                    "Sensors"
                ],
                "summary": "Obtener un sensor por ID",
                "operationId": "841b8aada35a85445909f9e1ba2914a5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del sensor",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sensor encontrado"
                    },
                    "404": {
                        "description": "Sensor no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/turns": {
            "get": {
                "tags": [
                    "Turns"
                ],
                "summary": "Obtener todos los turnos",
                "operationId": "78fd976a6bf5135be62284e00adf23ae",
                "responses": {
                    "200": {
                        "description": "Lista de turnos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Turns"
                ],
                "summary": "Crear un nuevo turno",
                "operationId": "c3fe5cfc2db101a3e130c143a1a03293",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "fermentation_id",
                                    "start_time"
                                ],
                                "properties": {
                                    "fermentation_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-06-14T10:00:00"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": null,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "fermentation_id": 1,
                                "start_time": "2024-06-14T10:00:00",
                                "end_time": null
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Turno creado exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/turns/{id}": {
            "get": {
                "tags": [
                    "Turns"
                ],
                "summary": "Obtener un turno por ID",
                "operationId": "1d8982373892a9c38118f54d7326a653",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del turno",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Turno encontrado"
                    },
                    "404": {
                        "description": "Turno no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}